Add parameter checks and unit tests for user

+++ src/api/user/user.repository.js
@@ -17,6 +17,9 @@ export function getUsers(queryCondition = {}, inclAvatars = false) {
 }
 
 export function getUser(id, inclAvatar = false) {
+    if (!id) {
+        return Promise.reject(new Error('user id is required'));
+    }
     return new Promise((resolve, reject) => {
         let query;
         if (inclAvatar) {
@@ -32,6 +35,9 @@ export function getUser(id, inclAvatar = false) {
 }
 
 export function lookupUserByEmail(email, inclAvatar = false) {
+    if (!email) {
+        return Promise.reject(new Error('email is required'));
+    }
     return new Promise((resolve, reject) => {
         let query;
         if (inclAvatar) {
@@ -47,6 +53,9 @@ export function lookupUserByEmail(email, inclAvatar = false) {
 }
 
 export function deleteUser(id) {
+    if (!id) {
+        return Promise.reject(new Error('user id is required'));
+    }
     return new Promise((resolve, reject) => {
         User.findByIdAndRemove(id).exec()
             .then(user => resolve(user))
@@ -55,6 +64,12 @@ export function deleteUser(id) {
 }
 
 export function updateUser(id, userData) {
+    if (!id) {
+        return Promise.reject(new Error('user id is required'));
+    }
+    if (!userData) {
+        return Promise.reject(new Error('userData is required'));
+    }
     return new Promise((resolve, reject) => {
         User.findById(id).exec()
             .then(user => {
@@ -70,6 +85,12 @@ export function updateUser(id, userData) {
 }
 
 export function uploadUserAvatar(id, file, deleteAfterUpload = true) {
+    if (!id) {
+        return Promise.reject(new Error('user id is required'));
+    }
+    if (!file) {
+        return Promise.reject(new Error('avatar file is required'));
+    }
     return new Promise((resolve, reject) => {
         let userPromise = getUser(id);
         let avatarPromise = userPromise.then(user => {
@@ -88,6 +109,9 @@ export function uploadUserAvatar(id, file, deleteAfterUpload = true) {
 }
 
 export function signUpUser(userData) {
+    if (!userData) {
+        return Promise.reject(new Error('user data is required'));
+    }
     let newUser = new User(userData);
     return newUser.save();
 }
